move state logic to another file
