+++ lib/mah/game/server.ex
@@ -2,82 +2,7 @@ defmodule Mah.Game.Server do
   use GenStateMachine, callback_mode: :handle_event_function
   require Logger
 
-  defmodule GameState do
-    defstruct id: nil,
-              players: [],
-              honba: 0,
-              round: 1,
-              tsumo_player_index: 0,
-              tsumohai: nil,
-              hands: %{},
-              yamahai: [],
-              rinshanhai: [],
-              wanpai: []
-
-    def new(id) do
-      %__MODULE__{id: id}
-    end
-
-    def haipai(game) do
-      if length(Enum.dedup(game.players)) == 4 do
-        {:ok, do_haipai(game)}
-      else
-        {:error, :not_enough_players}
-      end
-    end
-
-    defp do_haipai(game = %__MODULE__{players: players}) do
-      %{players: players, hands: hands, yamahai: yamahai, rinshanhai: rinshanhai, wanpai: wanpai} = Mah.Mahjong.haipai(players, %{})
-
-      %__MODULE__{game | players: players, hands: hands, yamahai: yamahai, rinshanhai: rinshanhai, wanpai: wanpai}
-    end
-
-    def tsumo(game) do
-      [tsumohai | yamahai] = game.yamahai
-
-      {:ok, %__MODULE__{game | tsumohai: tsumohai, yamahai: yamahai}}
-    end
-
-    def tsumogiri(game = %__MODULE__{players: players, hands: hands, tsumo_player_index: tsumo_player_index, tsumohai: tsumohai}, player_id) do
-      if player_id == players[tsumo_player_index] do
-        hands =
-          Map.update!(hands, player_id, fn hand = %{sutehai: sutehai} ->
-            Map.put(hand, :sutehai, [%{hai: tsumohai, tsumogiri: true} | sutehai])
-          end)
-
-        # TODO: check can anyone furo
-
-        next_tsumo_player_index = rem(tsumo_player_index + 1, length(players))
-        game = %GameState{game | tsumohai: nil, tsumo_player_index: next_tsumo_player_index, hands: hands}
-
-        {:ok, game}
-      else
-        {:error, :not_your_turn}
-      end
-    end
-
-    def dahai(game = %__MODULE__{players: players, hands: hands, tsumo_player_index: tsumo_player_index, tsumohai: tsumohai}, player_id, dahai) do
-      if player_id == players[tsumo_player_index] do
-        if dahai in get_in(hands, [player_id, :tehai]) do
-          hands =
-            Map.update!(hands, player_id, fn hand = %{tehai: tehai, sutehai: sutehai} ->
-              hand
-              |> Map.put(:sutehai, [%{hai: dahai, tsumogiri: false} | sutehai])
-              |> Map.put(:tehai, [tsumohai | Enum.reject(tehai, &(&1 == dahai))])
-            end)
-
-          next_tsumo_player_index = rem(tsumo_player_index + 1, length(players))
-          game = %GameState{game | tsumohai: nil, tsumo_player_index: next_tsumo_player_index, hands: hands}
-
-          {:ok, game}
-        else
-          {:error, :not_in_your_hand}
-        end
-      else
-        {:error, :not_your_turn}
-      end
-    end
-  end
+  alias Mah.Game.State, as: GameState
 
   def child_spec(id) do
     %{id: id, start: {__MODULE__, :start_link, [id]}}
